Move ScmResolver to GitHubSCMFacade and fix some minor problems.
